+++ server/controllers/capi.js
@@ -28,7 +28,7 @@ module.exports = function(req, res, next) {
 	var articleV1Promise;
 	if (res.locals.flags.articleCapiV1Fallback) {
 		articleV1Promise = api.contentLegacy({
-				uuid: req.params[0],
+				uuid: req.params.id,
 				useElasticSearch: res.locals.flags.elasticSearchItemGet
 			})
 				// Some things aren't in CAPI v1 (e.g. FastFT)
@@ -44,7 +44,7 @@ module.exports = function(req, res, next) {
 	}
 
 	var articleV2Promise = api.content({
-		uuid: req.params[0],
+		uuid: req.params.id,
 		type: 'Article',
 		metadata: true,
 		useElasticSearch: res.locals.flags.elasticSearchItemGet
@@ -155,7 +155,7 @@ module.exports = function(req, res, next) {
 		.catch(function(err) {
 			if (err instanceof fetchres.BadServerResponseError) {
 				api.contentLegacy({
-						uuid: req.params[0],
+						uuid: req.params.id,
 						useElasticSearch: res.locals.flags.elasticSearchItemGet
 					})
 						.then(function(data) {
