--- views/more-on-topic.html
@@ -1,26 +1,0 @@
-<div class="more-on more-on--topic more-on--{{topic.taxonomy}}{{#if hasMainImage}} more-on--with-image{{/if}}">
-	<h2 class="more-on__title">{{topic.title}}
-		<a class="more-on__topic-name ng-title-link" href="{{topic.url}}" data-trackable="theme">{{topic.name}}</a>
-	</h2>
-	<div class="more-on__content">
-		{{#if topic.isAuthor}}
-			<img class="more-on__author-image" src="http://image.webservices.ft.com/v1/images/raw/http%3A%2F%2Fnext.ft.com%2Fassets%2Ficons%2Fprofile.svg?format=svg&tint=%23000000,%23000000&source=next.ft.com" width="100" height="100" />
-		{{/if}}
-		<ul class="more-on__articles ng-list-reset">
-			{{#each articles}}
-				<li class="more-on__article">
-					<a href='/{{id}}' class='more-on__article__link ng-title-link' data-trackable='headline'>
-						{{#if image}}
-							<img class="more-on__article__image" src="{{image}}" />
-						{{/if}}
-						{{{title}}}
-					</a>
-					<time class="more-on__article__timestamp ng-meta o-date" data-o-component="o-date" datetime="{{#dateformat}}{{publishedDate}}{{/dateformat}}" data-o-date-js>
-						{{#dateformat "dddd, d mmmm, yyyy"}}{{publishedDate}}{{/dateformat}}
-					</time>
-				</li>
-			{{/each}}
-		</ul>
-		<a class="more-on__all-link ng-title-link" href="{{topic.url}}" data-trackable="see all">See all</a>
-	</div>
-</div>
